// src/app/models/constante.ts
import { Estado } from './enums/estado';
import { Tipo } from './tipo';

export class Constante {

  v_valcons: string;
  v_nomcons: string;
  n_idconssupe: number;
  n_idcons: number;
  idconstante: number;
  v_descons: string;
  v_abrecons: string;
  n_discons: number;

  constructor() {
    this.v_valcons = '';
    this.v_nomcons = '';
    this.n_idconssupe = 0;
    this.n_idcons = 0;
    this.idconstante = 0;
    this.v_descons = '';
  }
}